Added possibility to create check without registered customer

+++ src/main/java/com/vinyl/dao/CustomerDao.java
@@ -18,5 +18,5 @@ public interface CustomerDao {
 
 	Short updateDiscount(int num, short discount);
 
-	Double getSumForAllPurchases(int num);
+	Integer getSumForAllPurchases(int num);
 }
